@@ -3,7 +3,7 @@ using static CommunityToolkit.Maui.Markup.GridRowsColumns;
 
 namespace HackerNews;
 
-class StoryDataTemplate : DataTemplate
+partial class StoryDataTemplate : DataTemplate
 {
 	public StoryDataTemplate() : base(CreateGrid)
 	{
@@ -18,7 +18,7 @@ class StoryDataTemplate : DataTemplate
 			(Row.Title, 20),
 			(Row.Description, 20),
 			(Row.BottomPadding, 1)),
-		
+
 		ColumnDefinitions = Columns.Define(
 			(Column.LeftPadding, 6),
 			(Column.Content, Star),
