@@ -19,10 +19,10 @@ public abstract class BaseAsyncValueCommand<TExecute, TCanExecute> : BaseCommand
 	/// Initializes a new instance of BaseAsyncValueCommand
 	/// </summary>
 	/// <param name="execute">The Function executed when Execute or ExecuteAsync is called. This does not check canExecute before executing and will execute even if canExecute is false</param>
-	/// <param name="canExecute">The Function that verifies whether or not AsyncCommand should execute.</param>
+	/// <param name="canExecute">The Function that verifies whether AsyncCommand should execute.</param>
 	/// <param name="onException">If an exception is thrown in the Task, <c>onException</c> will execute. If onException is null, the exception will be re-thrown</param>
 	/// <param name="continueOnCapturedContext">If set to <c>true</c> continue on captured context; this will ensure that the Synchronization Context returns to the calling thread. If set to <c>false</c> continue on a different context; this will allow the Synchronization Context to continue on a different thread</param>
-	protected private BaseAsyncValueCommand(Func<TExecute?, ValueTask>? execute,
+	private protected BaseAsyncValueCommand(Func<TExecute?, ValueTask>? execute,
 												Func<TCanExecute?, bool>? canExecute,
 												Action<Exception>? onException,
 												bool continueOnCapturedContext) : base(canExecute)
