@@ -120,9 +120,6 @@ class Tests_AsyncCommand : BaseAsyncCommandTest
 		AsyncCommand command = new AsyncCommand(NoParameterTask, commandCanExecute);
 		command.CanExecuteChanged += handleCanExecuteChanged;
 
-		void handleCanExecuteChanged(object? sender, EventArgs e) => didCanExecuteChangeFire = true;
-		bool commandCanExecute(object? parameter) => canCommandExecute;
-
 
 		Assert.That(command.CanExecute(null), Is.False);
 
@@ -146,5 +143,41 @@ class Tests_AsyncCommand : BaseAsyncCommandTest
 			Assert.That(didCanExecuteChangeFire, Is.True);
 			Assert.That(command.CanExecute(null), Is.True);
 		});
+
+		void handleCanExecuteChanged(object? sender, EventArgs e) => didCanExecuteChangeFire = true;
+
+		bool commandCanExecute(object? parameter) => canCommandExecute;
+	}
+
+	[Test]
+	public void AsyncCommand_ExecuteAsync_ExceptionHandling_Test()
+	{
+		//Arrange
+		var command = new AsyncCommand(NoParameterImmediateNullReferenceExceptionTask, onException: HandleException);
+		Exception? caughtException = null;
+
+		//Act
+		caughtException = Assert.ThrowsAsync<NullReferenceException>(() => command.ExecuteAsync());
+
+		//Assert
+		Assert.That(caughtException, Is.Not.Null);
+
+		void HandleException(Exception ex) => caughtException = ex;
+	}
+
+	[Test]
+	public void AsyncCommand_ExecuteAsync_ExceptionHandlingWithParameter_Test()
+	{
+		//Arrange
+		AsyncCommand<int> command = new AsyncCommand<int>(ParameterImmediateNullReferenceExceptionTask, onException: HandleException);
+		Exception? caughtException = null;
+
+		//Act
+		caughtException = Assert.ThrowsAsync<NullReferenceException>(() => command.ExecuteAsync(0));
+
+		//Assert
+		Assert.That(caughtException, Is.Not.Null);
+
+		void HandleException(Exception ex) => caughtException = ex;
 	}
 }
