@@ -14,6 +14,12 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 		const string stringEventArg = "Test";
 		bool didEventFire = false;
 
+		//Act
+		TestStringWeakEventManager.RaiseEvent(this, stringEventArg, nameof(TestStringEvent));
+
+		//Assert
+		Assert.That(didEventFire, Is.True);
+
 		void HandleTestEvent(object? sender, string? e)
 		{
 			if (sender is null || e is null)
@@ -31,12 +37,6 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 			didEventFire = true;
 			TestStringEvent -= HandleTestEvent;
 		}
-
-		//Act
-		TestStringWeakEventManager.RaiseEvent(this, stringEventArg, nameof(TestStringEvent));
-
-		//Assert
-		Assert.That(didEventFire, Is.True);
 	}
 
 	[Test]
@@ -48,6 +48,12 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 		const string stringEventArg = "Test";
 
 		bool didEventFire = false;
+
+		//Act
+		TestStringWeakEventManager.RaiseEvent(null, stringEventArg, nameof(TestStringEvent));
+
+		//Assert
+		Assert.That(didEventFire, Is.True);
 
 		void HandleTestEvent(object? sender, string e)
 		{
@@ -62,12 +68,6 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 			didEventFire = true;
 			TestStringEvent -= HandleTestEvent;
 		}
-
-		//Act
-		TestStringWeakEventManager.RaiseEvent(null, stringEventArg, nameof(TestStringEvent));
-
-		//Assert
-		Assert.That(didEventFire, Is.True);
 	}
 
 	[Test]
@@ -111,14 +111,69 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 
 		bool didEventFire = false;
 
-		void HandleTestEvent(object? sender, string e) => didEventFire = true;
-
 		//Act
 		TestStringWeakEventManager.RaiseEvent(this, "Test", nameof(TestEvent));
 
 		//Assert
 		Assert.That(didEventFire, Is.False);
 		TestStringEvent -= HandleTestEvent;
+
+		void HandleTestEvent(object? sender, string e) => didEventFire = true;
+	}
+
+	[Test]
+	public void WeakEventManagerTEventArgs_RaiseEvent_WithParameters()
+	{
+		//Arrange
+		TestStringEvent += HandleTestEvent;
+		bool didEventFire = false;
+
+		//Act
+		TestStringWeakEventManager.RaiseEvent(this, "Test", nameof(TestStringEvent));
+
+		//Assert
+		Assert.That(didEventFire, Is.True);
+
+		void HandleTestEvent(object? sender, string e)
+		{
+			if (sender is null)
+				throw new ArgumentNullException(nameof(sender));
+
+			Assert.Multiple(() =>
+			{
+				Assert.That(sender, Is.Not.Null);
+				Assert.That(sender.GetType(), Is.EqualTo(this.GetType()));
+
+				Assert.That(e, Is.Not.Null);
+			});
+
+			didEventFire = true;
+			TestStringEvent -= HandleTestEvent;
+		}
+	}
+
+	[Test]
+	public void WeakEventManagerTEventArgs_ExceptionHandling()
+	{
+		//Arrange
+		TestStringEvent += HandleTestEvent;
+		Exception? caughtException = null;
+
+		//Act
+		try
+		{
+			TestStringWeakEventManager.RaiseEvent(this, "Test", nameof(TestStringEvent));
+		}
+		catch (Exception ex)
+		{
+			caughtException = ex;
+		}
+
+		//Assert
+		Assert.That(caughtException, Is.Not.Null);
+		TestStringEvent -= HandleTestEvent;
+
+		void HandleTestEvent(object? sender, string e) => throw new NullReferenceException();
 	}
 
 	[Test]
@@ -163,13 +218,14 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 
 		TestStringEvent += HandleTestEvent;
 		TestStringEvent -= HandleTestEvent;
-		void HandleTestEvent(object? sender, string e) => didEventFire = true;
 
 		//Act
 		TestStringWeakEventManager.RaiseEvent(this, "Test", nameof(TestStringEvent));
 
 		//Assert
 		Assert.That(didEventFire, Is.False);
+
+		void HandleTestEvent(object? sender, string e) => didEventFire = true;
 	}
 
 	[Test]
@@ -272,8 +328,6 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 		TestStringEvent += HandleTestStringEvent;
 		bool didEventFire = false;
 
-		void HandleTestStringEvent(object? sender, string e) => didEventFire = true;
-
 		//Act
 
 		//Assert
@@ -284,5 +338,7 @@ class Tests_WeakEventManager_EventHandlerT : BaseTest
 		});
 
 		TestStringEvent -= HandleTestStringEvent;
+
+		void HandleTestStringEvent(object? sender, string e) => didEventFire = true;
 	}
 }
