@@ -6,15 +6,11 @@ namespace AsyncAwaitBestPractices;
 /// <summary>
 /// Represents errors that occur during WeakEventManager.HandleEvent execution.
 /// </summary>
-public class InvalidHandleEventException : Exception
+/// <remarks>
+/// Initializes a new instance of the <see cref="T:AsyncAwaitBestPractices.InvalidHandleEventException"/> class.
+/// </remarks>
+/// <param name="message">Message.</param>
+/// <param name="targetParameterCountException">Target parameter count exception.</param>
+public class InvalidHandleEventException(string message, TargetParameterCountException targetParameterCountException) : Exception(message, targetParameterCountException)
 {
-	/// <summary>
-	/// Initializes a new instance of the <see cref="T:AsyncAwaitBestPractices.InvalidHandleEventException"/> class.
-	/// </summary>
-	/// <param name="message">Message.</param>
-	/// <param name="targetParameterCountException">Target parameter count exception.</param>
-	public InvalidHandleEventException(string message, TargetParameterCountException targetParameterCountException) : base(message, targetParameterCountException)
-	{
-
-	}
 }
