@@ -3,14 +3,8 @@ using System.Reflection;
 
 namespace AsyncAwaitBestPractices;
 
-struct Subscription
+readonly struct Subscription(in WeakReference? subscriber, in MethodInfo handler)
 {
-	public Subscription(in WeakReference? subscriber, in MethodInfo handler)
-	{
-		Subscriber = subscriber;
-		Handler = handler ?? throw new ArgumentNullException(nameof(handler));
-	}
-
-	public WeakReference? Subscriber { get; }
-	public MethodInfo Handler { get; }
+	public WeakReference? Subscriber { get; } = subscriber;
+	public MethodInfo Handler { get; } = handler ?? throw new ArgumentNullException(nameof(handler));
 }
